+++ taktik/core/social_media/instagram/actions/atomic/text/text_input.py
@@ -5,7 +5,6 @@ from typing import Optional
 from loguru import logger
 
 from ...core.base_action import BaseAction
-from ....ui.selectors import TEXT_INPUT_SELECTORS, DETECTION_SELECTORS
 
 
 class TextInputMixin(BaseAction):